deconstructed todo to become individual components
