+++ src/components/todoComponent.vue
@@ -1,0 +1,120 @@
+<template>
+    <div class="column col-md-3 ">
+      <h1>todo</h1>
+      <div class="card" style="width: 18rem;" v-for="(todoItem, key) in todo"
+        v-bind:key=key>
+        <div class="card-body">
+          <h6 class="card-title">{{todoItem.todoTitle}}</h6>
+          <h6 class="card-subtitle mb-2 text-muted">code:{{todoItem.todoCode}}</h6>
+          <p class="card-text">{{todoItem.todoDescription}}</p>
+          <button @click="revertTodo(todoItem.todoCode, todoItem.todoTitle, todoItem.todoDescription)" type="button" class="btn btn-danger btn-sm">revert</button> | | 
+          <button @click="moveTodo(todoItem.todoCode, todoItem.todoTitle, todoItem.todoDescription)" type="button" class="btn btn-success btn-sm">progress</button>
+        </div>
+      </div>
+    </div>
+</template>
+
+<script>
+
+import { mapState , mapActions } from 'vuex';
+
+export default {
+    name: 'HelloWorld',
+  data(){
+    return {
+      msg : "welcome to hontoni subarashi kanban",
+      title: '',
+      description: ''
+    }
+  },
+  computed: {
+    ...mapState([
+      'todo',
+      'doing',
+      'done',
+      'backlog'
+    ])
+  }
+  ,
+  methods: {
+    ...mapActions([
+      'addBacklog',
+      'addTodo',
+      'getTodo',
+      'getDone',
+      'getBacklog',
+      'progressTodo',
+      'progressBacklog',
+      'getDoing',
+      'regressDoing',
+      'proceedDoing',
+      'regressDone',
+      'regressTodo',
+      'deleteDone'
+
+    ]),
+    submitBacklog(title, description){
+      let contentObj = {
+        title,
+        description
+      }
+      this.addBacklog(contentObj)
+    },
+    proceedBacklog(backlogCode, backlogTitle, backlogDescription){
+       let backlogObj = {
+         backlogCode,
+         backlogTitle,
+         backlogDescription 
+       };
+       this.progressBacklog(backlogObj)
+     },
+    revertTodo(todoCode, todoTitle, todoDescription){
+      let todoObj = {
+        todoCode,
+        todoTitle,
+        todoDescription 
+      };
+      this.regressTodo(todoObj)
+    },
+    moveTodo( todoCode, todoTitle, todoDescription ){
+      let todoObj = {
+        todoCode,
+        todoTitle,
+        todoDescription 
+      };
+      this.progressTodo(todoObj)
+    },
+    revertDoing(doingCode, doingTitle, doingDescription){
+      let doingObj = {
+        doingCode,
+        doingTitle,
+        doingDescription
+      };
+     this.regressDoing(doingObj)
+    },
+    progressDoing(doingCode, doingTitle, doingDescription){
+      let doingObj = {
+        doingCode,
+        doingTitle,
+        doingDescription
+      };
+      this.proceedDoing(doingObj)
+    },
+    revertDone(doneCode, doneTitle, doneDescription){
+      let doneObj = {
+        doneCode,
+        doneTitle,
+        doneDescription
+      }
+      this.regressDone(doneObj)
+    },
+    finishAndDelete(doneCode){
+      this.deleteDone(doneCode)
+    }
+   },
+    mounted(){
+      this.getTodo()
+    }
+}
+</script>
+
