+++ src/views/Home.vue
@@ -1,18 +1,15 @@
 <template>
   <div class="home">
-    <hello-world/>
-    
+    <h1>nasi goreng itu enak</h1>
   </div>
 </template>
 
 <script>
 // @ is an alias to /src
-import HelloWorld from '@/components/HelloWorld.vue'
 
 export default {
   name: 'home',
   components: {
-    HelloWorld
   }
 }
 </script>
